feat: Apply Multi Module
