+++ art-api/src/main/kotlin/com/flab/artshare/config/security/JwtAuthFilter.kt
@@ -1,6 +1,6 @@
 package com.flab.artshare.config.security
 
-import com.flab.artshare.AuthService
+import com.flab.artshare.firebase.FirebaseService
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.filter.OncePerRequestFilter
@@ -8,7 +8,7 @@ import javax.servlet.FilterChain
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
-class JwtAuthFilter(private val authService: AuthService) : OncePerRequestFilter() {
+class JwtAuthFilter(private val firbaseService: FirebaseService) : OncePerRequestFilter() {
     override fun doFilterInternal(
         request: HttpServletRequest,
         response: HttpServletResponse,
@@ -22,7 +22,7 @@ class JwtAuthFilter(private val authService: AuthService) : OncePerRequestFilter
 
     private fun updateAuthentication(token: String) {
         SecurityContextHolder.getContext().authentication =
-            UsernamePasswordAuthenticationToken(this.authService.getUserId(token), null, null)
+            UsernamePasswordAuthenticationToken(this.firbaseService.verifyToken(token), null, null)
     }
 
     private fun extractAuthToken(request: HttpServletRequest): String? =
