--- art-domain/src/main/kotlin/com/flab/artshare/AuthService.kt
@@ -1,9 +1,0 @@
-package com.flab.artshare
-
-import com.flab.artshare.firebase.FirebaseService
-import org.springframework.stereotype.Service
-
-@Service
-class AuthService(private val firebaseService: FirebaseService) {
-    fun getUserId(token: String): String = this.firebaseService.verifyToken(token)
-}
